requirements.txt
pvporcupine>=3.0
sounddevice>=0.4
numpy>=1.24
faster-whisper>=1.0
rapidfuzz>=3.0
pyyaml>=6.0
